mejoras en código, Readme y limpieza

+++ src/api/controllers/users.js
@@ -11,14 +11,17 @@ async function register(req, res) {
       return res.status(400).json('El usuario ya existe')
     }
     const newUser = new User({
-      username: req.body.username,
+      username,
       password: req.body.password,
       photo: req.file ? req.file.path : null
     })
-    await newUser.save()
-    return res.status(201).json(`Usuario creado correctamente: ${newUser}`)
+    const userSaved = await newUser.save()
+    return res.status(201).json(`Usuario creado correctamente: ${userSaved}`)
   } catch (error) {
     console.log('Error al crear usuario', error)
+    if (req.file) {
+      deleteImgCloud(req.file.path)
+    }
     return res.status(500).json(`Error al crear usuario`)
   }
 }
@@ -54,6 +57,9 @@ const getUser = async (req, res) => {
   try {
     const { id } = req.params
     const user = await User.findById(id).populate('images')
+    if (!user) {
+      return res.status(404).json('Usuario NO encontrado')
+    }
     return res.status(200).json(user)
   } catch (error) {
     console.log(`Error al obtener las imágenes por usuario: ${error}`)
@@ -63,18 +69,30 @@ const getUser = async (req, res) => {
 const updateUser = async (req, res) => {
   try {
     const { id } = req.params
-    const { username, password, photo, images } = req.body
-    const user = await User.findById(id)
+    if (!id) {
+      return res.status(404).json('Indique usuario')
+    }
+    const actualUser = await User.findById(id)
+    if (!actualUser) {
+      return res.status(404).json('Usuario NO encontrado')
+    }
+    const oldPath = actualUser.photo
+    if (req.file && oldPath !== req.file.path) {
+      deleteImgCloud(oldPath)
+    }
+    const { username, password, images } = req.body
     const newData = {
-      username: username || user.username,
+      username: username || actualUser.username,
       password: password
         ? bcrypt.hashSync(req.body.password, 10)
-        : user.password,
-      photo: req.file ? req.file.path : user.photo
+        : actualUser.password,
+      photo: req.file ? req.file.path : oldPath
     }
     let userImages = []
-    if (images) {
-      userImages = [...user.images, ...images]
+    if (Array.isArray(images)) {
+      userImages = [...actualUser.images, ...images]
+    } else {
+      userImages = [...actualUser.images, images]
     }
     const modifyUser = await User.findByIdAndUpdate(
       id,
@@ -86,7 +104,9 @@ const updateUser = async (req, res) => {
     return res.status(200).json(modifyUser)
   } catch (error) {
     console.log(`Error al actualizar el usuario: ${error}`)
-    deleteImgCloud(req.file.path)
+    if (req.file) {
+      deleteImgCloud(req.file.path)
+    }
     return res.status(500).json('Error al actualizar el usuario')
   }
 }
@@ -94,6 +114,9 @@ const deleteUser = async (req, res) => {
   try {
     const user = req.user
     const userDeleted = await User.findByIdAndDelete(user._id)
+    if (!userDeleted) {
+      return res.status(404).json('Usuario NO encontrado')
+    }
     if (userDeleted.photo) {
       deleteImgCloud(userDeleted.photo)
     }
